Changes were added to the map

+++ crud.py
@@ -42,7 +42,6 @@ def create_book(google_book_id, title, authors, poster_path, rating, number_of_r
         google_book_id = google_book_id,
         title=title,
         authors = ", ".join(authors) if authors is not None else "",
-        # authors = authors,
         poster_path=poster_path,
         rating = rating,
         number_of_ratings = number_of_ratings
@@ -116,10 +115,11 @@ def update_avg_rating(google_book_id, new_score):
         book.rating = float(new_score)
         book.number_of_ratings = 1
     else:
-        book.rating = (book.rating * book.number_of_ratings + int(new_score)) / (book.number_of_ratings + 1)
+        book.rating = float((book.rating * book.number_of_ratings + int(new_score)) / (book.number_of_ratings + 1))
         book.number_of_ratings += 1
 
         print(f"/////////  BOOK RATING: {book.rating} ///////////")
+
     return book.rating
 
 def create_bookgenre(book_id, genre_id):
